fix(nemesis): move traffic control helpers from AwsNode to BaseNode
Move install_traffic_control() and traffic_control() to BaseNode as they existed
only on AwsNode, so BlockNetworkMonkey and RandomInterruptionNetworkMonkey failed with
AttributeError on GCE once multi-NIC support made them eligible there.

Renamed install_traffic_control() to has_traffic_control(), since it only checks
for tc and never installs it. Comment notes this assumes tc ships on the image
(true for SMI-based AMIs/GCE images); the caller nemesis skips otherwise.

Refs: SCT-308
(cherry picked from commit 4614c803890b1f1bb39dc19c57682694dbe61774)
